README: Add secretbox example
I found it difficult to figure out how to get started with this and had to
spelunk through the code to figure out how the API for the libsodium wrappers
varied from the API for the underlying libsodium library (and the error
messages were opaque when I got it wrong).

Add an example for cryptobox of how to read a key, encrypt and decrypt a
message. Hopefully this will make it a little more clear about how to use this
API (and so people can apply these techniques to other libraries!) The example
text is largely taken from the secretbox example reviewed by Adam Langley,
here: https://go-review.googlesource.com/#/c/26810/.